python/projects/video-subtitiles/backend/audio_extraction/requirements.txt
google-cloud-storage==2.7.0
google-cloud-pubsub==2.13.0